Classes Scripting & Prefabs
Created a prefab for each class with its animator, script (empty script for now), basic sprite and collisions. From now on only the scripts should change when wanting to edit the classes (outside of projectiles like the elemental balls or the arrows).

Note: This commit might not work, as the player controller (base) script is missing.