Elementalist Scripting
Elementalist now casts fire / ice on left / right clicks, and during the cast animation the ball goes on to its way for a period of time (or until collides with something from the UI).
Mana decreases on use, and refreshes after few seconds.
Added animators for fire and ice, but currently the sprites for ice are pretty lame (fire with color change).
Also changed the health and mana pngs.