Warlord Scripting
Firstly, a block (custom made) animation was added, with the last minute decision to add this instead of right click that attacks as well. While animation is running, the player won't get hit by anything.
Secondly, the left click of the warlord was added as well, checking for enemies to hit, and also having an indicator to show the warlord's range on the editor.
Lastly, added a basic enemy script, just to have the functions.