Dialog
Added dialog to the game, now there are some instructions and story texts that the player sees once it reaches a certain point. The text stops the player from doing anything.
The dialog uses text mesh pro as a newly added asset.

Also mana refresh had a bug where if casting, then waiting enough time the next cast would be free because the refresh time things it is time to refresh.